requirements.txt
numpy==1.9.2
Pillow==2.9.0
scipy==0.15.1
